requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pytest>=7.4.0
pydantic>=2.4.0
openai>=1.3.0